feat: package restructure with CLI
