+++ tests/test_logger.py
@@ -1,11 +1,6 @@
 import logging
-import sys
-from pathlib import Path
 
-ROOT_DIR = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT_DIR))
-
-from logger import get_logger  # noqa: E402
+from web2pdfbook.logger import get_logger
 
 
 def test_get_logger_reuses_instance():
