+++ web2pdfbook/__init__.py
@@ -1,0 +1,1 @@
+"""web2pdfbook package."""
