+++ web2pdfbook/crawler/entity/crawl_result.py
